+++ src/ui/InternetWizard.tsx
@@ -5,9 +5,8 @@ import { Box, Text, useInput } from 'ink'
 import TextInput from './TextInput'
 import { SEARCH_PROVIDERS, activeSearchProvider } from '../config'
 import { t } from '../i18n'
+import { INDIGO, SILVER } from './theme'
 
-const INDIGO = '#6A5ACD'
-const SILVER = '#C8D8FF'
 const DIM = '#7A8AAA'
 const GREEN = '#5AF78E'